Now we can actually move!
Phu that was tough, first no picture at all since sizes in to
glDataBuffer where wrong.  After that stilil no move due to not running
glEnableVertexAttribPointer inside rednder loop which ment that
glDrawArrayInstanced silently fails. Added that but didnt notice that
one of the n * sizeif was wrong ( 2,2,3 insteas of 1*,2*3*)..
But now all works! Next step is to add color change in shaders from z
depth.